mttn: move syscall types into own module
